requirements.txt
astrapy==0.7.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]
openai==1.8.0
tiktoken==0.5.2
# langchain-openai==0.0.3
huggingface_hub
langchain
langchain-groq
langchain_community
